tuesday-feb

+++ app_forms/account_fees.py
@@ -1,21 +1,21 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, TextAreaField, FloatField, DateField, SubmitField
+from wtforms import StringField, TextAreaField, DateField, SubmitField, IntegerField
 from wtforms.validators import DataRequired,Length 
 
 class App_Account_Fees(FlaskForm):
-    deposits = FloatField('deposits', validators=[DataRequired()])
-    withdrawals = FloatField('withdrawals',validators=[DataRequired()] )
-    paycodes = FloatField('paycodes', validators=[DataRequired()])
-    payments = FloatField('payments', validators=[DataRequired()])
+    deposits = IntegerField('deposits', validators=[DataRequired()])
+    withdrawals = IntegerField('withdrawals',validators=[DataRequired()] )
+    paycodes = IntegerField('paycodes', validators=[DataRequired()])
+    payments = IntegerField('payments', validators=[DataRequired()])
     submit = SubmitField('Done')
 
 class Account_limits(FlaskForm):
-    withdrawal = FloatField('withdrawal', validators=[DataRequired()] )
-    paycode = FloatField('paycode', validators=[DataRequired()])
-    payments = FloatField('payments', validators=[DataRequired()])
-    daily = FloatField('daily', validators=[DataRequired()])
-    weekly = FloatField('weekly', validators=[DataRequired()])
-    monthly = FloatField('monthly', validators=[DataRequired()])
+    withdrawal = IntegerField('withdrawal', validators=[DataRequired()] )
+    paycode = IntegerField('paycode', validators=[DataRequired()])
+    payments = IntegerField('payments', validators=[DataRequired()])
+    daily = IntegerField('daily', validators=[DataRequired()])
+    weekly = IntegerField('weekly', validators=[DataRequired()])
+    monthly = IntegerField('monthly', validators=[DataRequired()])
     submit = SubmitField('Done')
